requirements.txt
biopython==1.79
certifi==2021.5.30
llvmlite==0.36.0
numba==0.53.1
numpy==1.21.1
pandas==1.3.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
wincertstore==0.2
